+++ src/Public/Add-ErRelation.ps1
@@ -1,4 +1,4 @@
-function Add-Relation {
+function Add-ErRelation {
 
     <#
 
@@ -15,10 +15,10 @@ function Add-Relation {
     None.
 
     .EXAMPLE
-    PS C:\> $diagram = New-MermaidDiagram -Type erDiagram
-    PS C:\> $diagram | Add-MermaidRelation Exactly-one Customer places Zero-or-more Order
-    PS C:\> $diagram | Add-MermaidRelation Exactly-one Order contains One-or-more LineItem
-    PS C:\> $diagram | Add-MermaidRelation One-or-more Customer uses One-or-more DeliveryAddress -NonIdentifying
+    PS C:\> $diagram = New-MermaidDiagram -ErDiagram
+    PS C:\> $diagram | Add-MermaidErRelation Exactly-one Customer places Zero-or-more Order
+    PS C:\> $diagram | Add-MermaidErRelation Exactly-one Order contains One-or-more LineItem
+    PS C:\> $diagram | Add-MermaidErRelation One-or-more Customer uses One-or-more DeliveryAddress -NonIdentifying
     PS C:\> $diagram | ConvertTo-MermaidString
     erDiagram
         Customer ||--o{ Order : places
@@ -36,6 +36,7 @@ function Add-Relation {
     param (
         # The diagram, that the relation is added to.
         [Parameter(ValueFromPipeline)]
+        [ValidateNotNull()]
         $Diagram,
 
         # The first entity of the relation.