+++ src/Public/Add-FlowchartLink.ps1
@@ -1,9 +1,10 @@
-function Add-Link {
+function Add-FlowchartLink {
 
     [CmdletBinding()]
     param (
-        # The diagram, that the relation is added to.
+        # The flowchart diagram, that the relation is added to.
         [Parameter(ValueFromPipeline)]
+        [ValidateNotNull()]
         $Diagram,
 
         # The source node of the link.