+++ src/Public/Add-FlowchartNode.ps1
@@ -1,9 +1,10 @@
-function Add-Node {
+function Add-FlowchartNode {
 
     [CmdletBinding()]
     param (
         # The diagram, that the relation is added to.
         [Parameter(ValueFromPipeline)]
+        [ValidateNotNull()]
         $Diagram,
 
         # The identifier of the node.