+++ test/Add-FlowchartNode.Tests.ps1
@@ -1,15 +1,19 @@
-Describe Add-Node {
+#Requires -Modules @{ ModuleName='Pester'; ModuleVersion='5.0.0' }
+
+Describe Add-FlowchartNode {
+
     BeforeAll {
         Import-Module $PSScriptRoot/../src/PsMermaidTools.psd1 -Force
     }
 
     Context flowchart {
+
         BeforeEach {
-            $diagram = New-MermaidDiagram -Type flowchart -Orientation top-down
+            $diagram = New-MermaidDiagram -Flowchart -Orientation top-down
         }
 
         It works-by-named-parmeters {
-            $diagram | Add-MermaidNode -Key A -Name foo -Shape round-edges
+            $diagram | Add-MermaidFlowchartNode -Key A -Name foo -Shape round-edges
             $diagram | Should -Not -BeNullOrEmpty
             $diagram.Nodes | Should -Not -BeNullOrEmpty
             $diagram.Nodes.Count | Should -Be 1
@@ -19,7 +23,7 @@ Describe Add-Node {
         }
 
         It works-by-positions {
-            $diagram | Add-MermaidNode A foo round-edges
+            $diagram | Add-MermaidFlowchartNode A foo round-edges
             $diagram | Should -Not -BeNullOrEmpty
             $diagram.Nodes | Should -Not -BeNullOrEmpty
             $diagram.Nodes.Count | Should -Be 1