+++ CHANGELOG.md
@@ -9,7 +9,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- new diagram type flowcharts.
+- new diagram type flowchart.
+- new command `Add-FlowchartNode`.
+- new command `Add-FlowchartMermaidLink`.
+- new diagram type C4 component diagram.
+- new command `Add-C4Component`.
+- new command `Add-C4ContainerBoundary`.
+- new command `Add-C4Relation`.
+- new command `New-C4Component`.
+- new command `New-C4ContainerBoundary`.
+
+### Changed
+
+- renamed `Add-Relation` to `Add-ErRelation`.
 
 ## [0.2.0] - 2022-09-14